feat!: remove db extension commands and unflag database primitive commands
