--- src/commands/database/legacy/constants.ts
@@ -1,3 +1,0 @@
-export const NEON_DATABASE_EXTENSION_SLUG = process.env.NEON_DATABASE_EXTENSION_SLUG ?? 'neon'
-export const JIGSAW_URL = process.env.JIGSAW_URL ?? 'https://jigsaw.services-prod.nsvcs.net'
-export const NETLIFY_NEON_PACKAGE_NAME = '@netlify/neon'
