+++ src/commands/database/util/packages.ts
@@ -1,6 +1,6 @@
 import { log } from '../../../utils/command-helpers.js'
 import BaseCommand from '../../base-command.js'
-import { spawnAsync } from '../legacy/utils.js'
+import { spawnAsync } from './spawn-async.js'
 
 export type PkgManagerName = 'npm' | 'yarn' | 'pnpm' | 'bun'
